requirements.txt
Flask
openai==0.28
python-dotenv

